first draft of new hochbeet firmware
