+++ platformio.ini
@@ -9,7 +9,28 @@
 ; https://docs.platformio.org/page/projectconf.html
 [platformio]
 default_envs = adafruit_feather_m0
-;default_envs = uno
+
+[env:adafruit_feather_m0]
+platform = atmelsam
+board = adafruit_feather_m0
+framework = arduino
+lib_deps =
+    Adafruit Unified Sensor
+    Adafruit BME280 Library
+    MCCI LoRaWAN LMIC library
+    VL6180X
+    RTCZero
+build_flags = 
+    -DDEBUG
+    -DBOARD_FEATHER_M0
+    -DARDUINO_LMIC_PROJECT_CONFIG_H_SUPPRESS
+    ; toggle for OTAA, if disabled, ABP is enabled
+    ;-DOTAA
+    -DCFG_eu868
+    -DCFG_sx1276_radio
+
+
+
 
 [env:megaatmega2560]
 platform = atmelavr
@@ -25,6 +46,7 @@ lib_deps =
     https://github.com/pololu/vl6180x-arduino
     RTClib
 build_flags =
+    -DBOARD_ARDUINO_MEGA
     -DTest
 ;    -DOLED
     -DDEBUG
@@ -70,28 +92,3 @@ build_flags =
 ;lib_ldf_mode = chain+
 
 
-[env:adafruit_feather_m0]
-platform = atmelsam
-board = adafruit_feather_m0
-framework = arduino
-lib_deps =
-    Adafruit Unified Sensor
-    Adafruit BME280 Library
-    Adafruit GFX Library    
-    Adafruit SSD1306
-    MCCI LoRaWAN LMIC library
-    https://github.com/pololu/vl6180x-arduino
-    RTCZero
-    GxEPD2
-build_flags = 
-    -DTest
-    -DDEBUG
-    -DDEBUG1
-    -DRTClock
-;    -DTTN
-    -DOTAA
-    -DCFG_eu868
-    -DPiBME280
-;    -DEPAPER
-
-;lib_ldf_mode = chain+
