+++ src/config/pinmaps/featherm0.h
@@ -1,0 +1,12 @@
+#ifndef FEATHER_PIN_MAP_H
+#define FEATHER_PIN_MAP_H
+
+// PIN MAP for Feather M0
+const lmic_pinmap lmic_pins = {
+    .nss = 8,
+    .rxtx = LMIC_UNUSED_PIN,
+    .rst = LMIC_UNUSED_PIN,
+    .dio = {3, 6, LMIC_UNUSED_PIN},
+};
+
+#endif
