+++ src/config/pins.h
@@ -1,0 +1,14 @@
+#ifndef PIN_CONFIG_H
+#define PIN_CONFIG_H
+
+
+#define TENSIOMETER_PRESSURE_PIN A1
+// Water Level Sensors
+//#define PIN_RELEASE_BUTTON 13   //Buttton to reset Error
+#define PIN_WATER_TANK_EMPTY 12 //Tank (A4)
+#define PIN_FLOWER_POT_FULL 13  //flowers
+
+#define PIN_RELAY A5
+
+
+#endif
